In ARM disassembler, print apparent fixnum-valued operands using QUOTE if there are GPR operands and all such operands are node regs. Use ~S when printing operands in some other cases.
In ARM backend, revive some changes that were backed out of a few
days ago and implement some similar things: extend the register-
tracking mechanism to deal with constants (t/nil/fixnums/immediates),
try to better integrate the register-tracking stuff with the things
that do "soft" register targeting (and deal with conflicts that may
be introduced as a result.)  Skip some function-call overhead on
some self-calls.

If we aren't bounds-checking and have a known-fixnum index in ARM2-VSET,
don't bother to load that index into a register.

The lisp builds itself and passes all tests with these changes in effect;
some of the issues in getting to that point are subtle so this should
be viewed with at least some suspicion.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14922 8bdb8f1d-59ce-db11-9e41-0016172a54ae