%HARD-REGSPEC-VALUE masks its argument by the platform-specific value of *ENCODED-REG-VALUE-BYTE*; the intent is that this allows us to be more consistent in referring to things like "the 0th FPR", regardless of how that's represented on a given platform.
Bind *ENCODED-REG-VALUE-BYTE* in compiler backends.

The temp-fp masks for x8632 and x8664 need to be built with this masking
in mind.

The CHECK-MIN-MAX-NARGS vinsns on x8632 and x8664 erroneously defined
MAX to be a :U16 (register) rather than a :U16CONST (constant).  The
vinsn machinery uses the %HARD-REGSPEC-VALUE of a register argument;
previously, that caused MAX to be treated (MOD 256) and it now causes
it to be (MOD 16) on x8664 and (MOD 8) on x8632, making this bug a
little easier to notice.

in arm-vinsns.lisp: remove a stray ^X character and note that the
SAVE-LISP-CONTEXT-OFFSET vinsn clobbers arm::imm0.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14959 8bdb8f1d-59ce-db11-9e41-0016172a54ae