Support keeping float-typed variable values in non-volatile FPRs on the ARM (so far.)
library/lispequ.lisp: move the definition of the VAR istruct from here
to compiler/nxenv.lisp so that changes to it affect the compiler when
the system is rebuilt.

compiler/nxenv.lisp: move the VAR istruct definition here; rename the
previously unused VAR-DECLARED-TYPE slot to VAR-DECLARED-UNBOXED-TYPE.
In NX-INIT-VAR, pay attention to TYPE declarations involving (for now)
SINGLE-FLOAT and DOUBLE-FLOAT and set the VAR-DECLARED-UNBOXED-TYPE of
the var accordingly.  (The idea is to recognize these types even if
the declaration or nested local declarations create a specialized subtype
like (AND DOUBLE-FLOAT (SATISFIES FOO)).)

compiler/nx0.lisp: define a PRINT-OBJECT method for VARs that at least
prints the VAR's name.  (I don't know if conflicting methods are defined
elsewhere.)

compiler/nx2.lisp: some changes to the global register allocator
(basically, to allow it to run on the same set of candidates more than
once per function).  Add NX2-SELECT-FPR-CANDIDATES, which considers
things that had their VAR-DECLARED-UNBOXED-TYPE set or which were bound
to an initial value of some floating-point type and were never SETQed.

compiler/vinsn.lisp: support SINGLE-FLOAT and DOUBLE-FLOAT temporary
registers with a specified value (mostly so that we can note when
specific registers of those types get clobbered.)  Support :ADDRESS-typed
temporaries.

compiler/vreg.lisp: when printing LREGs, try to show their mode (mostly
to make it easier to distinguish between SINGLE- and DOUBLE-FLOAT regs
when printing vinsns.)

compiler/ARM/arm-asm.lisp: don't set the W (writeback) bit in the definition
of fldmias/fldmiad (it can be set by an operand.)  Allow instructions that
accept a (general-purpose) register list operand to specify the list via
(:$ ,bitmask) where bitmask is a 16-bit integer.

compiler/ARM/arm-vinsns.lisp: new vinsns to save/restore non-volatile FPRs.
Try to avoid moving between FPRs and GPRs in more cases.  (We still do that
to initialize floats to constant values; if we can't figure out FCONST[S|D]
or can't use it, we might want to load float constants from memory.)

compiler/ARM/arm2.lisp: exploit the above to allow the ARM backend to
use non-volatile FPRs.  We may want to tweak this a bit: we now avoid
float-consing in cases where we used to do a lot of it and now do some
in other cases where we didn't used to.  In general, if a float-typed
variable is SETQed, keeping it in an FPR usually wins (because we can
often avoid consing in that case.)  If such a variable isn't SETQed, then
keeping it in an FPR can cons if the variable's returned from the defining
function or passed as an argument in a function call.  In cases where the
value doesn't change, keeping both the boxed and unboxed representations
may be worth considering.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14972 8bdb8f1d-59ce-db11-9e41-0016172a54ae