Add (%nx1-operator div2) to *numeric-acode-ops*; this allows us to do some simple type inference/handle some cases of numeric contagion involving two-arg #'/.  (We don't do anything with INTEGER-typed args; if we ever did, we'd have to be careful with #'/ here.)
In NX1-NOTE-VAR-BINDING, be careful: NX-ACODE-FORM-TYPE can
destrictively modify its argument, making an acode from that's had
type info stripped from it bcome a type-asserted form.  Change the
order in which the variables INIT and INITTYPE are initialized so
that we can recognize:

{{{
  (let* ((v0 v1))
    ...)
}}}

and consider whether or not the two variables need to be disjoint.

While we're at it: ACODE-FORM-TYPE (which does these destructive type
assertions) is supposed to plead ignorance when it encounters a
lexical reference and it's called from the frontend.  It really
shouldn't assert that ignorance, which may inhibit subsequent calls
from the backend from determining type info when more information
is available.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14980 8bdb8f1d-59ce-db11-9e41-0016172a54ae