We've traditionally tried to estimate the number of references/assignments to a VAR by using two 8-bit fields in its VAR-BITS slot.  That means that we have to treat 255 references/assignments as being infinite (rather than "unusually large"); we're also out of bits on 32-bit machines. Stop using these bitfields; use two new slots in the VAR istruct instead. We've been inconsisten about whether or not the bitfields were loop-weighted or not.  They aren't, but the VAR-REFS slot in the per-function version of a VAR is.
Change frontend sources to stop using the bitfields; add accessors/setters
for the new slots and use them instead.  Like NX-VAR-BITS, these slots are
only maintained for "root" (non-inherited) VARs.

in compiler/*/???2.lisp: in the function that tries to eliminate &REST args,
there was never any need to test the (old) $vsetqmask field.  Since that 
field doesn't exist anymore, stop doing so.

Add IMM-REG-P to backend.lisp.

in compiler/ARM/arm2.lisp:

ARM2-CR-BIT-FOR-REVERSED-COMPARISON: handle unsigned comparisons too.

ARM2-TWO-UNTARGETED-REG-FORMS: if both target registers are imm regs
and both expressions look "simple", mark the first used imm reg as
being in use before evaluating the second form.  (Some cases of
evaluating "simple" things to imm regs may otherwise want to use
other imm regs as temporaries.)

ARM2-CONSTANT-FOR-COMPARE-P takes an optional UNBOXED argument.

Add ARM2-NATURAL-COMPARE, which does comparisons between (UNSIGNED-BYTE 32)s.
Make ARM2-%NATURAL<> use it.

If we process a binding of A to B and conclude that B is so uninteresting
that A doesn't need to exist, if the VAR for A has type information about
B replace refs to A with the acode equivalent of (THE <type> B).

This all seems to work on ARM; haven't tried to compile on other platforms.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@14982 8bdb8f1d-59ce-db11-9e41-0016172a54ae