vinsn.lisp: add two new slots to the VINSN struct, to track GPRs and FPRs read by the VINSN.  Deal (crudely) with bootstrapping issues that this change raises (vreg.lisp loads earlier and there are circular dependencies there.)  Add code to access these new slots, similar to that which had existed to track register assignments.  When looking to see if a vinsn sequence references a register, return the last vinsn which does so (not the first.)  When testing attributes in a vinsn sequence in %VINSN-SEQUENCE-HAS-ATTRIBUTE-P, logand the vinsn's attributes with the supplied "attr" arg, not with nothing: we want to see if a vinsn has all of the supplied attributes, not test for an exact match.
vreg.lisp: MATCH-VREG updates the new -refs slots in the vinsn.  More
circularity and bootstrapping workarounds.

backend.lisp: %AVAILABLE-NODE-TEMP returns register number or NIL;
AVAILABLE-NODE-TEMP calls it and errs on null return.

x8632-vinsns.lisp: catch a few more cases where we can't access the low
byte of high-numbered GPRs.  (Who designed this cpu ?)  Add a few
vinsns for comparing fixnums to stack locations and doing fixnum
addition on stack locations.  Make sure that MARK-AS-NODE notes that
the argument register is modified.

x8664-vinsns.lisp: comparisons/additions on stack locations.

x862.lisp: a bunch of changes to try to reduce stack traffic a little
more.  Do some cases of fixnum comparison/addition directly on stack
locations if the location's value isn't known to be in a GPR.  (This
may be faster than a load or load+store and may reduce register
pressure a bit.  We do this if the stack location isn't in a register;
the real question is whether we want it to be, possibly more than we
want something that's currently in a register to stay there.)  When
loading a set of values into an arbitrary set of registers
(x862-*-untargeted-reg-forms), try to use an unused temporary if the
suggested register contains a stack location value.  (Again, the real
question is whether or not it contains a value that we'd prefer keeping
there.)  This change means that vinsns generally operate on a wider set
of registers than they had previously and exposed some issues (extracting
the low byte of high-numbered registers on x8632, avoiding conflicts when
some registers have dedicated roles due to hardware (shifts by variable
amounts) or software (consing uses allocptr/temp0).
X862-ELIDE-PUSHES tries to remove speculative pushes/pops on the vstack.
In order for this to work, we have to know whether or not the vinsn sequence
between PUSH Rx and POP Ry references Ry and whether the last such reference
precedes the first assignment (if any) to Rx.  (This is why we needed to
track register references in vinsns.  A reference to Ry can occur in that
sequence becuause Ry just happens to contain a cached stack location or
other interesting value.)

When these functions (and their cousins, the -targeted-reg-forms
functions) think that they need to push a register, the heuristics
that they use to try to decide between pushing a specific register or
an arbitrary one have changed (and aren't always consistent.)  We want
to avoid copies before the push, but also want to elide the push if
possible and want to avoid introducing another copy if we do so.
It's more important to be able to eliminate push/pop pairs than it
is to eliminate a register copy or two, so we should probably think
in those terms.

This stuff is a work in progress; it builds and passes its test suite
on Linuxx8632/64, but it wouldn't be a total shock if bugs were intoduced.
I think that we see fewer load, pushes, and pops (which was the goal)
and more register copies (some of which are probably necessary and some
of which are just kind of silly-looking ...)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15022 8bdb8f1d-59ce-db11-9e41-0016172a54ae