When allocating "untargeted" temporary registers, heuristically try to avoid conflicts with registers that may contain live values.  These heuristics occasionally lose, but often win and lead to a reduction in stack traffic for many functions.
"elide" some PUSHes/POPs on the vstack.  Since LOADs and STOREs to
the vstack are VSP-relative, adust the appropriate operand of any
intervening VFRAME-LOAD and VFRAME-STORE vinsns if we're able to
elide the push/pop.

Fix a few things that implicitly and incorrectly assumed that fewer
node temps were used (e.g., that assumed that their result was being
stored in arm::arg_z or that the source and destination operands of
a vinsn were the same.)

While bootstrapping these changes, I ran into mysterious cache coherency
problems.  I was never able to resolve them (among other things, running
under GDB made the problems go away.)  It's not at all clear how compiler
changes could cause these sorts of problems, but backing out of those 
changes and incrementally reintroducing them made the problems go away.

(I saw the problems on a TrimSlice and a Tegra develpment board, both
of which are dual-core Cortex A9 (Nvidia Tegra) systems.  I tried a
few different Linux kernels on the TrimSlice but didn't try other ARM
boxes.  My best guess is that this is an SMP issue and that some stale
cache state is retaine on one CPU when the other CPU invalidates it.
and that I just happened to be able to provoke a CPU bug.)

Whatever the problem is, I'll try to check in an ARM heap image that reflects
thes changes and those in the next commit; if the problem recurs, reverting
to that image may un-wedge things.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15066 8bdb8f1d-59ce-db11-9e41-0016172a54ae