ARM-specific changes, mostly.
When running lisp code (in almost all cases), keep the constant 0.0d0
in the d7 register (and therefore 0.0s0 in s14 and s15).  We use d7 as
a vector header when saving non-volatile FPRs on the stack; we
actually only modify s14, so we now restore s14 after it's been used
this way.  The value used in the header in lisp and kernel code is
loaded from PC-relative memory, which means that we no longer use 
fmsr/fmdrr or similar instructions.

When starting a lisp thread or entering one via a callback, initialize
d7.

This all basically means that we can get 0.0[d|s]0 into an FPR (or
exploit the fact that it's already in one) a bit easier, and that's
generally a good thing.  It's an ABI change, which means that the
FASL and image versions (for the ARM port only) changed; new binaries
are included in this commit.

The kernel changes to support the use of d7 are mostly pretty obvious.
In working on them, I noticed that "local labels" and "macro labels"
were in the same namespace, and we were only avoiding conflicts by
accident.  For 10 years or so.  (I also noticed that GAS doesn't fully
support PC-relative operands, so did that by hand.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15111 8bdb8f1d-59ce-db11-9e41-0016172a54ae