Change the initial-values *TERMINAL-CHARACTER-ENCODING-NAME* and *DEFAULT-FILE-CHARACTER-ENCODING* to :UTF-8, mostly for the benefit of the Init-File-Editing-Impaired.  (I've resolved not to make fun of the IFEI.)  Note that this may require changes to startup scripts etc.
Define new conditions CCL:DECODING-PROBLEM and CCL:ENCODING-PROBLEM.
Signal these conditions (via SIGNAL) when decoding characters
from/enoding them to a stream, pointer or octet vector and a substitution
or replacement character would be used.

New macros (CCL:WITH-DECODING-PROBLEMS-AS-ERRORS &body body) and
(CCL:WITH-ENCODING-PROBLEMS-AS-ERRORS &body body) signal the corresponding
conditions as ERRORs if they are signaled during execution of the body.

(Arguably) fixes ticket:749.

FILE-STRING-LENGTH checks to see if the encoding wants to use a
byte-order-mark before subtracting the length of an encoded BOM
from the encoded string length if the file is at its beginning.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15236 8bdb8f1d-59ce-db11-9e41-0016172a54ae