cocoa-ide/app-delegate.lisp: move MAYBE-FIXUP-APPLICATION-MENU here, split it off from INITIALIZE-USER-INTERFACE, call it in delegate's #/applicationWillFinishLaunching: method
cf-utils.lisp: don't depend on read-time behavior of #_ in macroexpansions.

cocoa-application.lisp: try to nuke command-line args when this file
is loaded (so that Cocoa doesn't try to process things like '-e
\(require \"COCOA-APPLICATION\"\)'.)

cocoa-backtrace.lisp: in COUNT-STACK-DESCRIPTOR-FRAME, only count
frames that satisfy CCL::FUNCTION-FRAME-P.  (I believe that this is
correct for some fairly subtle reason that I don't remember ...)

cocoa-editor.lisp,cocoa-listener.lisp: better support multiple views/windows
on the same buffer.  (The listener support should more-or-less work, but
there are secondary listener-specific issues.

Add an item to a Hemlock text view's context menu which offers to duplicate
the current window; that creates a new window/view hierarchy that shares
the text view's buffer and document but maintains view-specific selection
state.  Windows and views created via this menu item are functionally
equvalent to the original; if the shared document is modified, the Cocoa
document architecture will offer to save it (on those OSX versions that
trust the appliance owner/user to make this sort of decision ...) when the
last window sharing the document is closed.  (The last window needn't be
the original.)  Listeners have state (streams) that refer to the original
window/view and closing that window/view before closing secondary windows
doesn't work; we need to either change how listener streams work or make
it hard to close the original window when duplicates of it still exist,
but I haven't done this.

The HEMLOCK-VIEW for a text view/echo-area view is maintained in a slot;
the idea's to make it a bit easier to embed multiple text views (which
may refer to multiple buffers) in a window (where the concept of "the
window's HEMLOCK-VIEW" may not apply.)

hemlock/src/command.lisp: a buffer's mark-ring is just a field referenced
by a BUFFER structure, not the value of a buffer-local Hemlock variable.

hemlock/src/killcoms.lisp: mark-ring changes, don't put any reasonable
number of consecutive ^Ds on the kill ring.

hemlock/src/lispmode.lisp:  don't loop forever at end/beginning of buffer
in %LIST-OFFSET macro.  Provide a terse description of the error in most
EDITOR-ERROR calls in this file.

hemlock/src/morecoms.lisp: mark-ring changes.

hemlock/src/struct.lisp: move some fields (-point, -%mark, the mark-ring,
the active region indicator) from BUFFER to a new SELECTION-INFO struct
that BUFFER references; redefine traditional accessors to indirect through
the SELECTION-INFO.  (Views can maintain their own SELECTION-INFO; view-based
operations temporarily install the view's SELECTION-INFO in the buffer.)

Add a TEXTSTORAGE slot to BUFFER; the idea (NYI) is that buffers can have
Cocoa textstorage associated with them independent of the buffer's DOCUMENT
object.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15291 8bdb8f1d-59ce-db11-9e41-0016172a54ae