Warn (via full warning) on duplicate bindings in SYMBOL-MACROLET, MACROLET, FLET, LABELS (fixes ticket:927.)
If a DECLARE expression is encountered when a form is expected,
make the error message more verbose (and mention macroexpansion
as a possible cause of the problem.)  Fixes ticket:926.

Warn (via a full warning) if a local function shadows a global
CL function name.  Fixes ticket:923.

If STYLE-WARNINGs are incidentally signaled during (e.g.)
macroexpansion and are handled and postprocessed by the compiler,
ensure that the warning actually generated will be a STYLE-WARNING.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15314 8bdb8f1d-59ce-db11-9e41-0016172a54ae