cocoa-ide/builder-utilites.lisp: in MAKE-DOCTYPE-DICT, declare unused keyword args to be ignorable.
cocoa-ide/cocoa-application.lisp, objc-bridge/objc-runtime.lisp: move
code which nukes NXArgc/NXArgv to a point just before Cocoa is loaded,
where it seems to work as intended.  This makes it safer than it has
been to (REQUIRE "COCOA") in a lisp that was started with command-line
arguments.

See ticket:955, and likely some others.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15327 8bdb8f1d-59ce-db11-9e41-0016172a54ae